Prettify JSON/Uglify Go
